Move many function to color.py in util
