+++ .env
@@ -1,4 +1,4 @@
 FLASK_ENV=development
 FLASK_DEBUG=1
 FLASK_APP=server_config.py
-MONGO_URI=MONGO_URI
+MONGO_URI=mongodb+srv://Fluk:1234@cluster0.gjhxa.mongodb.net/test
